Disable production-mode minimization
From this diff, it may appear that only CSS minification is being disabled here; but in fact the 'minimize' flag also controls webpack's default JavaScript minimization approach.  We are disabling mimization of both CSS and JavaScript, and bypassing the use of TerserPlugin

In production mode, minimization was consuming the bulk of bundling time; ~30 seconds out of a ~45 second build.  That probably is recovered on aggregate for large-enough deployment bases with small-enough deployment schedules; in our case at the moment it seems better to optimize for fewer development-time dependencies, closer similarity between development and production, and faster production deployment times