Use plain, unprocessed template index.html file when running unit tests -- skipping webpack stage
There's a chance that this will break at some point; we're relying on the fact that the template index.html is 'close enough' to the output HTML that unit tests will still be able to find relevant DOM elements and modify them as needed.  Generally that _should_ be the case - all we're doing at the moment during the webpack build is adding external 'link' and 'script' references to the compiled application modules.