Add a 'Referrer Policy', using an HTML meta tag
Currently this should be somewhat redundant, because all of the non-homepage bookmarkable/navigable URLs used within the application are based on URL fragment content (information included after the '#' character in the URL).  Even so, it's possible that will change in future - particularly if we enable progressive enhancement support - so let's add this as a precaution.

Ref: https://www.w3.org/TR/referrer-policy/#referrer-policy-origin
Ref: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Referrer-Policy#integration_with_html